pyproject.toml
[tool.poetry]
name = "syncman"
version = "1.0.0"
description = "Simple Sync Manager"
authors = ["bissli <bissli@xyz.com>"]

[tool.poetry.dependencies]
python = "~3.11"

more-itertools = "^10.1.0"
cachetools = "^5.3.1"
python-dateutil = "^2.8.2"
pytz = "^2023.3.post1"
piny = "^1.1.0" # not sure needed

psycopg = { extras = ["binary", "pool"], version = "^3.1.12" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.22" }
pandas = "^2.1.2"
pytest-lazy-fixture = "^0.6.3"

[tool.poetry.group.dev.dependencies]
pytest = { extras = ["aysncio"], version = "^7.4.2" }
asserts = "^0.12.0"
docker = "^6.1.3"
pytest-postgresql = "^5.0.0"
# pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
